consolidate resources into one
without a single hierarchy getting the cache/adapter types is awkward as it would shadow the GET {idOrName} call
