+++ graylog2-server/src/main/java/org/graylog2/rest/resources/system/lookup/LookupTableResource.java
@@ -5,16 +5,24 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonUnwrapped;
 
 import org.apache.shiro.authz.annotation.RequiresAuthentication;
+import org.graylog2.lookup.MongoLutCacheService;
+import org.graylog2.lookup.MongoLutDataAdapterService;
 import org.graylog2.lookup.MongoLutService;
+import org.graylog2.lookup.dto.CacheDto;
+import org.graylog2.lookup.dto.DataAdapterDto;
 import org.graylog2.lookup.dto.LookupTableDto;
 import org.graylog2.rest.models.PaginatedList;
+import org.graylog2.rest.models.system.lookup.CacheApi;
+import org.graylog2.rest.models.system.lookup.DataAdapterApi;
 import org.graylog2.rest.models.system.lookup.LookupTableApi;
 import org.graylog2.shared.rest.resources.RestResource;
 import org.hibernate.validator.constraints.NotEmpty;
 import org.mongojack.DBQuery;
 import org.mongojack.DBSort;
 
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import javax.annotation.Nullable;
@@ -23,7 +31,9 @@ import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.DefaultValue;
 import javax.ws.rs.GET;
+import javax.ws.rs.NotFoundException;
 import javax.ws.rs.POST;
+import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
@@ -34,20 +44,27 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 
 @RequiresAuthentication
-@Path("/system/lookup/tables")
+@Path("/system/lookup")
 @Produces("application/json")
 @Consumes("application/json")
-@Api(value = "System/Lookup/Tables", description = "Lookup tables")
+@Api(value = "System/Lookup", description = "Lookup tables")
 public class LookupTableResource extends RestResource {
 
-    private MongoLutService lookupTableService;
+    private final MongoLutService lookupTableService;
+    private final MongoLutDataAdapterService adapterService;
+    private final MongoLutCacheService cacheService;
 
     @Inject
-    public LookupTableResource(MongoLutService lookupTableService) {
+    public LookupTableResource(MongoLutService lookupTableService,
+                               MongoLutDataAdapterService adapterService,
+                               MongoLutCacheService cacheService) {
         this.lookupTableService = lookupTableService;
+        this.adapterService = adapterService;
+        this.cacheService = cacheService;
     }
 
     @GET
+    @Path("tables")
     @ApiOperation(value = "List configured lookup tables")
     public LookupTablePage tables(@ApiParam(name = "page") @QueryParam("page") @DefaultValue("1") int page,
                                   @ApiParam(name = "per_page") @QueryParam("per_page") @DefaultValue("50") int perPage,
@@ -79,6 +96,7 @@ public class LookupTableResource extends RestResource {
     }
 
     @POST
+    @Path("tables")
     @ApiOperation(value = "Create a new lookup table")
     public LookupTableApi createTable(@ApiParam LookupTableApi lookupTable) {
         LookupTableDto saved = lookupTableService.save(lookupTable.toDto());
@@ -86,7 +104,7 @@ public class LookupTableResource extends RestResource {
     }
 
     @DELETE
-    @Path("{idOrName}")
+    @Path("tables/{idOrName}")
     @ApiOperation(value = "Delete the lookup table")
     public void removeTable(@ApiParam(name = "idOrName") @PathParam("idOrName") @NotEmpty String idOrName) {
         // TODO validate that table isn't in use
@@ -110,6 +128,166 @@ public class LookupTableResource extends RestResource {
             this.query = query;
             this.paginationInfo = paginationInfo;
             this.lookupTables = lookupTables;
+        }
+    }
+
+    @GET
+    @Path("adapters")
+    @ApiOperation(value = "List available data adapters")
+    public DataAdapterPage adapters(@ApiParam(name = "page") @QueryParam("page") @DefaultValue("1") int page,
+                                @ApiParam(name = "per_page") @QueryParam("per_page") @DefaultValue("50") int perPage,
+                                @ApiParam(name = "sort",
+                                        value = "The field to sort the result on",
+                                        required = true,
+                                        allowableValues = "title")
+                                @DefaultValue("title") @QueryParam("sort") String sort,
+                                @ApiParam(name = "order", value = "The sort direction", allowableValues = "asc, desc")
+                                @DefaultValue("desc") @QueryParam("order") String order,
+                                @ApiParam(name = "query") @QueryParam("query") String query) {
+
+        PaginatedList<DataAdapterDto> paginated = adapterService.findPaginated(DBQuery.empty(), DBSort.asc(sort), page, perPage);
+        return new DataAdapterPage(query,
+                paginated.pagination(),
+                paginated.stream().map(DataAdapterApi::fromDto).collect(Collectors.toList()));
+    }
+
+    @GET
+    @Path("types/adapters")
+    @ApiOperation(value = "List available data adapter types")
+    public List<String> availableAdapterTypes() {
+        return Collections.emptyList();
+    }
+
+    @GET
+    @Path("adapters/{idOrName}")
+    @ApiOperation(value = "List the given data adapter")
+    public DataAdapterApi getAdapter(@ApiParam(name = "idOrName") @PathParam("idOrName") @NotEmpty String idOrName) {
+        Optional<DataAdapterDto> dataAdapterDto = adapterService.get(idOrName);
+        if (dataAdapterDto.isPresent()) {
+            return DataAdapterApi.fromDto(dataAdapterDto.get());
+        }
+        throw new NotFoundException();
+    }
+
+    @POST
+    @Path("adapters")
+    @ApiOperation(value = "Create a new data adapter")
+    public DataAdapterApi createAdapter(@ApiParam DataAdapterApi newAdapter) {
+        DataAdapterDto dto = newAdapter.toDto();
+        DataAdapterDto saved = adapterService.save(dto);
+        return DataAdapterApi.fromDto(saved);
+    }
+
+    @DELETE
+    @Path("adapters/{idOrName}")
+    @ApiOperation(value = "Delete the given data adapter", notes = "The data adapter cannot be in use by any lookup table, otherwise the request will fail.")
+    public void deleteAdapter(@ApiParam(name = "idOrName") @PathParam("idOrName") @NotEmpty String idOrName) {
+        // TODO validate that adapter isn't in use
+        adapterService.delete(idOrName);
+    }
+
+    @PUT
+    @Path("adapters")
+    @ApiOperation(value = "Update the given data adapter settings")
+    public DataAdapterApi updateAdapter(@ApiParam DataAdapterApi toUpdate) {
+        DataAdapterDto saved = adapterService.save(toUpdate.toDto());
+        return DataAdapterApi.fromDto(saved);
+    }
+
+    @JsonAutoDetect
+    public static class DataAdapterPage {
+        @Nullable
+        @JsonProperty
+        private final String query;
+
+        @JsonUnwrapped
+        private final PaginatedList.PaginationInfo paginationInfo;
+
+        @JsonProperty("data_adapters")
+        private final List<DataAdapterApi> dataAdapters;
+
+        public DataAdapterPage(@Nullable String query, PaginatedList.PaginationInfo paginationInfo, List<DataAdapterApi> dataAdapters) {
+            this.query = query;
+            this.paginationInfo = paginationInfo;
+            this.dataAdapters = dataAdapters;
+        }
+    }
+
+    @GET
+    @Path("caches")
+    @ApiOperation(value = "List available caches")
+    public CachesPage caches(@ApiParam(name = "page") @QueryParam("page") @DefaultValue("1") int page,
+                             @ApiParam(name = "per_page") @QueryParam("per_page") @DefaultValue("50") int perPage,
+                             @ApiParam(name = "sort",
+                                     value = "The field to sort the result on",
+                                     required = true,
+                                     allowableValues = "title")
+                             @DefaultValue("created_at") @QueryParam("sort") String sort,
+                             @ApiParam(name = "order", value = "The sort direction", allowableValues = "asc, desc")
+                             @DefaultValue("desc") @QueryParam("order") String order,
+                             @ApiParam(name = "query") @QueryParam("query") String query) {
+        PaginatedList<CacheDto> paginated = cacheService.findPaginated(DBQuery.empty(), DBSort.asc(sort), page, perPage);
+        return new CachesPage(query,
+                paginated.pagination(),
+                paginated.stream().map(CacheApi::fromDto).collect(Collectors.toList()));
+    }
+
+    @GET
+    @Path("types/caches")
+    @ApiOperation(value = "List available caches types")
+    public List<String> availableCacheTypes() {
+        return Collections.emptyList();
+    }
+
+    @GET
+    @Path("caches/{idOrName}")
+    @ApiOperation(value = "List the given cache")
+    public CacheApi getCache(@ApiParam(name = "idOrName") @PathParam("idOrName") @NotEmpty String idOrName) {
+        Optional<CacheDto> cacheDto = cacheService.get(idOrName);
+        if (cacheDto.isPresent()) {
+            return CacheApi.fromDto(cacheDto.get());
+        }
+        throw new NotFoundException();
+    }
+
+    @POST
+    @Path("caches")
+    @ApiOperation(value = "Create a new cache")
+    public CacheApi createCache(@ApiParam CacheApi newCache) {
+        return CacheApi.fromDto(cacheService.save(newCache.toDto()));
+    }
+
+    @DELETE
+    @Path("caches/{idOrName}")
+    @ApiOperation(value = "Delete the given cache", notes = "The cache cannot be in use by any lookup table, otherwise the request will fail.")
+    public void deleteCache(@ApiParam(name = "idOrName") @PathParam("idOrName") @NotEmpty String idOrName) {
+        // TODO validate that cache isn't in use
+        cacheService.delete(idOrName);
+    }
+
+    @PUT
+    @Path("caches")
+    @ApiOperation(value = "Update the given cache settings")
+    public CacheApi updateCache(@ApiParam CacheApi toUpdate) {
+        return CacheApi.fromDto(cacheService.save(toUpdate.toDto()));
+    }
+
+    @JsonAutoDetect
+    public static class CachesPage {
+        @Nullable
+        @JsonProperty
+        private final String query;
+
+        @JsonUnwrapped
+        private final PaginatedList.PaginationInfo paginationInfo;
+
+        @JsonProperty("caches")
+        private final List<CacheApi> caches;
+
+        public CachesPage(@Nullable String query, PaginatedList.PaginationInfo paginationInfo, List<CacheApi> caches) {
+            this.query = query;
+            this.paginationInfo = paginationInfo;
+            this.caches = caches;
         }
     }
 }
